requirements.txt
psycopg2-binary
flask
python-dotenv
